minor changes + added zustand persistence for players
