+++ src/features/schedule/types/training-block.type.ts
